Define output as c_float.

+++ src/fortran_routing/mc_pylink_v00/MC_singleSeg_singleTS/pyMCsingleSegStime_NoLoop.f90
@@ -13,7 +13,7 @@ subroutine c_muskingcungenwm(dt, qup, quc, qdp, ql, dx, bw, tw, twcc,&
     real(c_float), intent(in) :: qup, quc, qdp, ql
     real(c_float), intent(in) :: dx, bw, tw, twcc, n, ncc, cs, s0
     real(c_float), intent(in) :: velp, depthp
-    real, intent(out) :: qdc, velc, depthc
+    real(c_float), intent(out) :: qdc, velc, depthc
 
     call muskingcungenwm(dt, qup, quc, qdp, ql, dx, bw, tw, twcc,&
     n, ncc, cs, s0, velp, depthp, qdc, velc, depthc)
